Init repo, read explanation :tada:
This is the initial commit for a fun demo project that calls data from
InsightVM, an extremely robust, cloud-based vulnerability scanner. This
was initially a React app (in a now-deleted repo), but since this app
depends on an API that:

- I only have access to through a 30-day free trial
- Has to run on my machine (and subsequently could only be served over
  ngrok or something in this case)
- Is really resource intensive

... I figured it'd probably be smart to migrate to Next.js and generate
a static version of things. A bittersweet turn of events - I was looking
forward to having an exclusive, limited-time, promotional demo app.